# swagger.json
{
  "swagger": "2.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Swagger testing for Admin/Sub-Admin",
    "description": "This is assignment of Admin/sub-Admin done by vaneet kumar",
    "contact": { "email": "vaneet.k@applify.co" },
    "license": {
      "name": "",
      "url": ""
    }
  },
  "schemes": ["http", "https"],
...
  "securityDefinitions": {
    "bearerAuth": {
      "name": "Authorization",
      "in": "header",
      "type": "apiKey",
      "description": "JWT Authorization header"
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/feedback/feed/addFeed": {
      "post": {
        "summary": "add feedback",
        "tags": ["Feedback"],
        "parameters": [
          {
            "name": "uId",
            "in": "formData",
            "description": "enter userId",
            "required": "true",
            "type": "string"
          },
          {
            "name": "comment",
            "in": "formData",
            "description": "enter your comment",
            "required": "true",
            "type": "string"
          },
          {
            "name": "rating",
            "in": "formData",
            "description": "give ratings",
            "enum": ["1 star", "2 star", "3 star", "4 star", "5 star"],
            "required": "true",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": ""
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/feedback/feed/view/{id}": {
      "get": {
        "summary": "getting feedback details",
        "tags": ["Feedback"],
        "description": "get all feedback details",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": "true",
            "description": "Enter particular feedback Id",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
    "/feedback/ratingAndComment": {
      "get": {
        "summary": "Filter on the based on rating and comment",
        "tags": ["Feedback"],
        "content": ["application/json"],
        "parameters": [
          {
            "name": "limit",
            "description": "Limit",
            "in": "query",
            "required": true,
            "type": "integer",
            "value": "10"
          },
          {
            "name": "skip",
            "description": "Skip",
            "in": "query",
            "required": true,
            "type": "integer",
            "value": "0"
          },
          {
            "name": "sortBy",
            "description": "sort by",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": ["CreatedAt", "comment", "rating"],
            "value": ""
          },
          {
            "name": "orderBy",
            "description": "order by",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": ["ASC", "DESC"],
            "value": ""
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
    "/feedback/feed/deleteFeed": {
      "delete": {
        "tags": ["Feedback"],
        "summary": "Delete Feedback",
        "description": "Admin can delete feedbacks",
        "content": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "formData",
            "description": "enter feedback id",
            "required": "true",
            "type": "string"
          },
          {
            "name": "adminId",
            "in": "formData",
            "description": "enter user-id who gave the particular feedback",
            "required": "true",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Deletion Successful",
            "schema": {
              "type": ""
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/category/addCategory": {
      "post": {
        "security": [{ "Bearer": [] }],
        "summary": "Will add the parent categories",
        "tags": ["Categories"],
        "content": ["application/json"],
        "parameters": [
          {
            "name": "name",
            "in": "formData",
            "required": "true",
            "type": "string"
          },
          {
            "name": "image",
            "in": "formData",
            "required": "true",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Category Added Successfully",
            "schema": {
              "type": "object"
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/category/list": {
      "get": {
        "tags": ["Categories"],
        "description": "List all the categories that are not deleted, alongwith count of categories. And also we can filter categories by name.",
        "parameters": [
          {
            "name": "limit",
            "description": "Limit",
            "in": "query",
            "required": true,
            "type": "integer",
            "value": "10"
          },
          {
            "name": "skip",
            "description": "Skip",
            "in": "query",
            "required": true,
            "type": "integer",
            "value": "0"
          },
          {
            "name": "sortBy",
            "description": "sort by",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": ["name", "cId", "createdAt", "image"],
            "value": ""
          },
          {
            "name": "orderBy",
            "description": "order by",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": ["ASC", "DESC"],
            "value": ""
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search",
            "type": "string",
            "value": ""
          }
        ],
        "responses": {}
      }
    },
    "/category/update/category": {
      "put": {
        "tags": ["Categories"],
        "summary": "Edit Or Update / Admin Can Change The Name Of The Parent Category",
        "content": ["application/json"],
        "parameters": [
          {
            "name": "cId",
            "in": "formData",
            "description": "Enter parent category id",
            "required": "true",
            "type": "string"
          },
          {
            "name": "name",
            "in": "formData",
            "description": "Enter New Name For The Category",
            "required": "true",
            "type": "string"
          },
          {
            "name": "image",
            "in": "formData",
            "required": "true",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Edit Or Updation Successful",
            "schema": {
              "type": "object"
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
...
    "/FAQ/add": {
      "post": {
        "summary": "Add all FAQ",
        "tags": ["FAQ"],
        "content": ["application/json"],
        "parameters": [
          {
            "name": "name",
            "in": "formData",
            "description": "Enter user name",
            "required": "true",
            "type": "string"
          },
          {
            "name": "email",
            "in": "formData",
            "description": "Enter user Email",
            "required": "true",
            "type": "string"
          },
          {
            "name": "FAQ",
            "in": "formData",
            "description": "Enter your question",
            "required": "true",
            "type": "string"
          },
          {
            "name": "status",
            "in": "formData",
            "description": "Enterstatus",
            "enum": ["solved", "unsolved"],
            "required": "true",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "FAQ and user Added Successfully",
            "schema": {
              "type": "object"
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/FAQ/filter": {
      "get": {
        "summary": "getting FAQ filter details",
        "tags": ["FAQ"],
        "description": "get FAQ details",
        "parameters": [
          {
            "name": "limit",
            "description": "Limit",
            "in": "query",
            "required": true,
            "type": "integer",
            "value": "10"
          },
          {
            "name": "skip",
            "description": "Skip",
            "in": "query",
            "required": true,
            "type": "integer",
            "value": "0"
          },
          {
            "name": "sortBy",
            "description": "sort by",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": ["userType", "CreatedAt"],
            "value": ""
          },
          {
            "name": "orderBy",
            "description": "order by",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": ["ASC", "DESC"],
            "value": ""
          },
          {
            "name": "name",
            "in": "formData",
            "description": "Enter user name",
            "required": false,
            "type": "string"
          },
          {
            "name": "status",
            "in": "query",
            "description": "Enter status",
            "enum": ["solved", "unsolved"],
            "required": false,
            "type": "string",
            "value": ""
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
...
    "/FAQ/edit": {
      "put": {
        "summary": "Edit user.Need to login first and need token",
        "tags": ["FAQ"],
        "content": ["application/json"],
        "parameters": [
          {
            "name": "Faqid ",
            "in": "formData",
            "description": "Enter Faqid ",
            "required": "true",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Edit Successfull",
            "schema": {
              "type": ""
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/FAQ/delete": {
      "delete": {
        "tags": ["FAQ"],
        "summary": "Delete FAQ. Need to login then need token",
        "description": "Admin can delete FAQ",
        "content": ["application/json"],
        "parameters": [
          {
            "name": "Faqid",
            "in": "formData",
            "description": "Enter feedback id",
            "required": "true",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Deletion Successful",
            "schema": {
              "type": ""
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
...
    "/api/delete/{id}": {
      "delete": {
        "summary": "Delete user.Need to login first and need token",
        "tags": ["User"],
        "content": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "formData",
            "description": "Enter User Id",
            "required": "true",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Delete Successfull",
            "schema": {
              "type": ""
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    }
  }
}